requirements.txt
requests==2.32.3
argparse==1.4.0
paho-mqtt==2.1.0